startup: enable the MMU
This patch is for the boot loader to remove the restriction on
the operating position. By enabling MMU, it becomes possible to
operate at any address. Since we have only ported it, we may make
changes later.

Address Mapping:
Virtual Address            ->     Physical Address
0xFE000000(text-base)      ->     Current Address (Ex:0x5F000000)

Text-base of BL32 is fixed to 0xFE000000. This address is used as
the virtual address. However, the physical address is different
from the address that the user wants to use. Therefore, map the
current address to 0xFE000000. See the reference documentation for
details.

Adds:
   - Enable the MMU (use the section description page table)

Change-Id: I9cc66723f31a3a68b72ae0bfe6451f49f5848c4c
Signed-off-by: DeokJin Lee <truevirtue@nexell.co.kr>