startup: enable the MMU
This patch is for the boot loader to remove the restriction on
the operating position. By enabling MMU, it becomes possible to
operate at any address. Since we have only ported it, we may make
changes later.

For example, assume that the current address is 0x5F000000.
The total usage area is 16M. And the code area is 1M.
The page table consists of a section table (1MB).

Address Mapping:
Virtual Address         ->	Phsical Address
0x00000000--0x5EFFFFFF	->	0x00000000--0x5EFFFFFF	Non-Cached
0x5F000000--0x5F0FFFFF  ->	0x5F000000--0x5F0FFFFF  Cached (CODE:TEXT AREA)
0x5F100000--0xFDFFFFFF	->	0x5F100000--0xFDFFFFFF  Non-Cached
0xFE000000--0xFEFFFFFF	->	0xFE000000--0xFEFFFFFF  Non-Cached
0xFF000000--0xFFEFFFFF  ->	0xFF000000--0xFFEFFFFF	Non-Cached
0xFFF00000--0xFFFFFFFF	->	0xFFF00000--0xFFFFFFFF  Cached (SRAM MEMORY)

Text-base of BL32 is fixed to 0xFE000000. This address is used as
the virtual address. However, the physical address is different
from the address that the user wants to use. Therefore, map the
current address to 0xFE000000. See the reference documentation for
details.

Adds:
   - Enable the MMU (use the section description page table)

Change-Id: I9cc66723f31a3a68b72ae0bfe6451f49f5848c4c
Signed-off-by: DeokJin Lee <truevirtue@nexell.co.kr>